requirements/base.txt
Django==3.2.5
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
django-environ==0.4.5
django-extensions==3.1.3
celery==5.1.2
redis==3.5.3
psycopg2-binary==2.9.1